+++ tests/unit/test_otel_span.py
@@ -6,9 +6,7 @@ from unittest import TestCase
 from unittest import main
 from unittest.mock import Mock
 
-from trulens.experimental.otel_tracing.core.instrument import (
-    _resolve_attributes,
-)
+from trulens.core.otel.instrument import _resolve_attributes
 from trulens.experimental.otel_tracing.core.span import (
     set_user_defined_attributes,
 )
